GenotypeBit change to float

+++ src/pgl/infra/dna/genotype/GenotypeTable.java
@@ -172,7 +172,7 @@ public interface GenotypeTable {
      * @param siteIndex
      * @return
      */
-    public double getSiteHeterozygoteFraction (int siteIndex);
+    public float getSiteHeterozygoteFraction (int siteIndex);
 
     /**
      * Return byte value of the minor allele
@@ -193,7 +193,7 @@ public interface GenotypeTable {
      * @param siteIndex
      * @return
      */
-    public double getMinorAlleleFrequency (int siteIndex);
+    public float getMinorAlleleFrequency (int siteIndex);
 
     /**
      * Return the byte value of the major allele
@@ -235,7 +235,7 @@ public interface GenotypeTable {
      * @param siteIndex
      * @return
      */
-    public double getReferenceAlleleFrequency (int siteIndex);
+    public float getReferenceAlleleFrequency (int siteIndex);
 
     /**
      * Return the byte value of the alternative allele
@@ -256,7 +256,7 @@ public interface GenotypeTable {
      * @param siteIndex
      * @return
      */
-    public double getAlternativeAlleleFrequency (int siteIndex);
+    public float getAlternativeAlleleFrequency (int siteIndex);
 
     /**
      * Return the start index of a chromosome, inclusive
@@ -276,9 +276,9 @@ public interface GenotypeTable {
      * Return the genetic divergence of two taxa, Dxy is defined as 1 - IBS (Identify by state)
      * @param taxonIndex1
      * @param taxonIndex2
-     * @return Double.NaN if no shared non-missing sites exist
+     * @return Float.NaN if no shared non-missing sites exist
      */
-    public double getDxy (int taxonIndex1, int taxonIndex2);
+    public float getDxy (int taxonIndex1, int taxonIndex2);
 
     /**
      * Return the genetic divergence of two taxa in a specified region, Dxy is defined as 1 - IBS (Identify by state)
@@ -286,38 +286,38 @@ public interface GenotypeTable {
      * @param taxonIndex2
      * @param startSiteIndex inclusive
      * @param endSiteIndex exclusive
-     * @return Double.NaN if no shared non-missing sites exist
+     * @return Float.NaN if no shared non-missing sites exist
      */
-    public double getDxy (int taxonIndex1, int taxonIndex2, int startSiteIndex, int endSiteIndex);
+    public float getDxy (int taxonIndex1, int taxonIndex2, int startSiteIndex, int endSiteIndex);
 
     /**
      * Return the genetic divergence of two taxa based on a list of sites, Dxy is defined as 1 - IBS (Identify by state)
      * @param taxonIndex1
      * @param taxonIndex2
      * @param siteIndices
-     * @return Double.NaN if no shared non-missing sites exist
+     * @return Float.NaN if no shared non-missing sites exist
      */
-    public double getDxy (int taxonIndex1, int taxonIndex2, int[] siteIndices);
+    public float getDxy (int taxonIndex1, int taxonIndex2, int[] siteIndices);
 
     /**
      * Return a matrix of genetic divergence of all taxa based on all sites
      * @return
      */
-    public double[][] getDxyMatrix ();
+    public float[][] getDxyMatrix ();
 
     /**
      * Return a matrix of genetic divergence of all taxa based on 10K evenly distributed sites
      * When total site number > 10K, size = 10K; or size = total site number
      * @return
      */
-    public double[][] getDxyMatrixFast10K ();
+    public float[][] getDxyMatrixFast10K ();
 
     /**
      * Return a matrix of genetic divergence of all taxa based selected sites
      * @param siteIndices
      * @return
      */
-    public double[][] getDxyMatrix (int[] siteIndices);
+    public float[][] getDxyMatrix (int[] siteIndices);
     /**
      * Return a genotype table by sub-setting the current table by sites
      * @param siteIndices
